Telnet framework functioning!
* Handles DO/DONT/WILL/WONT negotiation of known Options.
* Handles IAC escapes.
* Passes all incoming data to all known (enabled) Options.
* Can send data from all (enabled) Options.

Basically, we treat the TelentClient like an object with State, and all Options maintain their own state. Output buffers are concatenated at the end of each read/write cycle. 

The consumeIncomingBytes() method gracefully handles incomplete frames. In the case that not enough data is present for frame decoding, bytes are redelivered in the next read/write loop.

Conceptually, this should work well enough for the beginning of a telnet client. Yay!





git-svn-id: https://jintn3270.svn.sourceforge.net/svnroot/jintn3270/trunk@20 d6bd4e80-8031-4e1d-bd39-93c407d940fc