Beginnings of AWT & Swing Terminal Widgets.
The following core ideas are going to be implemented by our widgets:
* Replaceable TerminalRenderers for doing the actual rendering of the TerminalModel.
  This will allow us to create different renderers for different terminal types, and potentially make it easy to extends the renderers with 
Options.
* All rendering occurrs in an off-screen buffer, which is rendered then repainted wholesale to the component.
* The TelnetFrame is a good test harness for the Terminal component.

Currently....
   The TerminalRenderer is responsible for reporting the minimum and preferred sizes of the Terminal component. It will report the preferred 
size based upon it's current Font setting, and the Minimum Size based upon the current font at 6 points.

We do not currently support scaling of fonts to the current size of the buffer, but we will in the future.

Next up is handling user input, sending it to the TerminalModel buffer (local echo) and out the TelnetClient output stream.





git-svn-id: https://jintn3270.svn.sourceforge.net/svnroot/jintn3270/trunk@39 d6bd4e80-8031-4e1d-bd39-93c407d940fc