Major refactoring....
Internally we now treat everything as 'shorts' to keep 0 ~ 255 values intact.

This addresses issues we've been having with java's lack of unsigned byte type. Basically, everything coming in from the socket 
is cast to a short with the value of (byte & 0xff). This is faster / better than if (byte < 0) byte +=256;

Yick.

Either way, it's a bit easier to code around now. As we can read shorts, return shorts, and let the TelnetClient handle 
transformation to / from byte[].






git-svn-id: https://jintn3270.svn.sourceforge.net/svnroot/jintn3270/trunk@80 d6bd4e80-8031-4e1d-bd39-93c407d940fc